Ahh shit, is a game actually being made here? Luv/Ppl Changes: * Changed default font * Talker boxes now have fucking backgrounds yeah * Implemented talker box closing * Implemented modular options for talker box * Additional tileset work * Tried to get transform gizmo on top of the world Engine 2D: * Added Box2D debug renderer * Fixed timestep issue on Box2D at occurred at over 100 FPS * Fixed bug in text width calculation * Increased autotile storage from 32 bytes to 64 bytes for special cases Renderer: * Added multiply and multiply x2 blend modes to material * Fixed null ref error in CTexture destructor on game end state * Reenabled double buffered primitives * Cleaned up debug line renderer's render Engine Common & Engine Base: * Transform::SetDirty and Transform::IsDirty will be removed * Changed interface for CRigidbody:   * No more auto-update via dirty transforms   * Added auto-update of only position   * Rigidbody can now be owned by CGameBehaviors as well as CGameObjects   * Rigidbody no longer need to have owners

+++ oneEngine-basic/source/engine-common/dusk/CDuskGUIHandle.cpp
@@ -6,7 +6,7 @@ CDuskGUI* Dusk::activeGUI = NULL;
 
 CDuskGUIElement* Dusk::Handle::operator*() const
 {
-	if (index >= 0 && index < activeGUI->vElements.size())
+	if (index >= 0 && index < (int)activeGUI->vElements.size())
 	{
 		return activeGUI->vElements[index];
 	}
@@ -14,7 +14,7 @@ CDuskGUIElement* Dusk::Handle::operator*() const
 }
 CDuskGUIElement* Dusk::Handle::operator->() const
 {
-	if (index >= 0 && index < activeGUI->vElements.size())
+	if (index >= 0 && index < (int)activeGUI->vElements.size())
 	{
 		return activeGUI->vElements[index];
 	}
